fetch_whether/src/requirements.txt
pandas
pandas-gbq
pyarrow
pytz
google-cloud-bigquery
google-cloud-logging
google-cloud-secret-manager